+++ CHANGELOG.md
@@ -16,6 +16,7 @@ FlatLaf Change Log
   `flatlaf-swingx-<version>.jar`.
 - CheckBox and RadioButton: Flag `opaque` is no longer ignored when checkbox or
   radio button is used as table cell renderer. (issue #77)
+- FileChooser: Use system icons. (issue #100)
 - FileChooser: Fixed missing labels in file chooser when running on Java 9 or
   later. (issue #98)
 - PasswordField: Do not apply minimum width if `columns` property is greater